requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
django-environ==0.11.2
gunicorn==21.2.0
Pillow==10.1.0
reportlab==4.0.7
numpy==1.26.2
scipy==1.11.4
geopy==2.4.1
pandas==2.1.3
plotly==5.18.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10